Make sm-server clippy and fmt gates pass on main (#1164) (#1529)
* Make sm-server clippy and fmt gates pass on main (#1164)

cargo clippy -p sm-server --all-targets -- -D warnings and
cargo fmt -p sm-server --check both exit 0. Code fixes throughout;
targeted too_many_arguments allows carry a one-line reason. AGENTS.md
tells every agent to run both before opening a PR.


* Drop a same-type cast main's #1535 added, keeping clippy clean


---------

+++ crates/sm-server/src/main.rs
@@ -319,7 +319,7 @@ fn raised_open_file_soft_limit(soft: u64, hard: u64, target: u64) -> Option<u64>
 mod tests {
     use super::raised_open_file_soft_limit;
 
-    const UNLIMITED: u64 = nix::libc::RLIM_INFINITY as u64;
+    const UNLIMITED: u64 = nix::libc::RLIM_INFINITY;
 
     #[test]
     fn raises_launchd_default_to_target_under_unlimited_hard_limit() {
